[*] Alot of improvements and more - read description
Fixed Wrong Object Pooling On Enemies, Fixed Projectile targeting new spawned enemies, added life and currency system, added GameOver and Pause Screen to game scene, and other UI setups for play scene, and alot of other improvements and fixes

+++ Petergramm/Assets/Editor/EnemyBalancingTool.cs
@@ -354,7 +354,7 @@ namespace Editor
             enemyName = selectedEnemy.enemyName;
             enemyPrefab = selectedEnemy.prefab;
 
-            hitpoints = selectedEnemy.stats.maxHp;
+            hitpoints = (int)selectedEnemy.stats.maxHp;
             damage = selectedEnemy.stats.damage;
             range = selectedEnemy.stats.range;
             attacksPerSecond = selectedEnemy.stats.attacksPerSecond;
